Rename file `jwt_generate` to `jwt`
